+++ server/main.py
@@ -1,4 +1,3 @@
-# server/main.py
 import asyncio
 from contextlib import asynccontextmanager
 from typing import List, Literal
